fix credential tests for upstream

+++ cypress/integration/tests/administrator/credentials/maven_crud.test.ts
@@ -27,6 +27,7 @@ describe("Validation of Source Control Credentials", () => {
     });
 
     after("Cleaning up", () => {
+        if (hasToBeSkipped("@tier2")) return;
         mavenCredsUsername.delete();
     });
 });
